Update patch set 1
Patch Set 1:

I disagree. If the overlay isn't there, then neither should the call. Yes, I know it doesn't hurt, but to me it looks bad, almost rushed. A piece of code that does nothing. And if someone decides to add an overlay to the device at some point in the future (especially if it doesn't have an overlay at all), them checking and adding the call to the overlay should be a part of it.

But if you want it to stay, I'll push up an amend shortly.

Patch-set: 1