Update patch set 1
Patch Set 1:

Ethan, this works as intended on m7vzw, but I don't think it should be limited to m7vzw. Would you be interested in seeing this HTCQualcommRIL migrate back to m7-common? The only other thing it handles is unsols, which should not cause any issues on the gsm variant.

Patch-set: 1